+++ packages/authhero/src/routes/universal-login/common.tsx
@@ -75,7 +75,6 @@ export async function initJSXRoute(
 
   if (!loginSession) {
     throw new HTTPException(400, { message: "Login session not found" });
-  } else if (loginSession.session_id && allowSession) {
   }
   ctx.set("loginSession", loginSession);